[libc++] Simplify std::sort helpers
These helpers were introduced in https://reviews.llvm.org/D118029
and then `sort` itself was rewritten in https://reviews.llvm.org/D122780
and the helpers simplified in https://reviews.llvm.org/D144478 .

This patch further simplifies the helpers. We apply D144478's
optimization (to not count the number of swaps) to `__sort3` everywhere
it's used except for the one place in `nth_element` that requires the
return value.